feat: ensureAuthorization 미들웨어로 구현현
