+++ src/controller/userController.js
@@ -2,30 +2,13 @@ const conn = require('../mariadb');
 const crypto = require('crypto');
 const jwt = require('jsonwebtoken');
 const { StatusCodes } = require('http-status-codes');
-const ensureAuthorization = require('../utils/auth');
 
 const getUser = (req, res) => {
-  const authorization = ensureAuthorization(req);
-
-  if (authorization instanceof ReferenceError) {
-    return res.status(StatusCodes.UNAUTHORIZED).json({
-      message: '로그인을 해주세요',
-    });
-  }
-  if (authorization instanceof jwt.JsonWebTokenError) {
-    return res.status(StatusCodes.BAD_REQUEST).json({
-      message: '잘못된 토큰입니다.',
-    });
-  }
-  if (authorization instanceof jwt.TokenExpiredError) {
-    return res.status(StatusCodes.UNAUTHORIZED).json({
-      message: '로그인 세션이 만료되었습니다. 다시 로그인 하세요',
-    });
-  }
+  const decodedJWT = req.decodedJWT;
 
   let sql = `SELECT id, name, email FROM pp_user 
     WHERE id = ?`;
-  let values = [authorization.id];
+  let values = [decodedJWT.id];
   conn.query(sql, values, (err, result) => {
     if (err) {
       return res.status(StatusCodes.BAD_REQUEST).json({
