+++ src/routes/users.js
@@ -1,3 +1,4 @@
+const ensureAuthorization = require('../middlewares/auth.js');
 const express = require('express');
 const router = express.Router();
 
@@ -8,7 +9,7 @@ const {
   editUser,
 } = require('../controller/userController.js');
 
-router.get('/', getUser);
+router.get('/', ensureAuthorization, getUser);
 router.post('/join', join);
 router.post('/login', login);
 router.put('/edit', editUser);
